pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "keydive"
version = "2.2.0"
description = "Extract Widevine L3 keys from Android devices effortlessly, spanning multiple Android versions for DRM research and education."
license = "MIT"
authors = ["hyugogirubato <65763543+hyugogirubato@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/hyugogirubato/KeyDive"
keywords = ["python", "drm", "widevine", "google"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
include = [
    { path = "CHANGELOG.md", format = "sdist" },
    { path = "README.md", format = "sdist" },
    { path = "LICENSE", format = "sdist" }
]

[tool.poetry.urls]
"Issues" = "https://github.com/hyugogirubato/KeyDive/issues"
"Packages" = "https://github.com/hyugogirubato/KeyDive/blob/main/docs/PACKAGE.md"
"Functions" = "https://github.com/hyugogirubato/KeyDive/blob/main/docs/FUNCTIONS.md"
"Changelog" = "https://github.com/hyugogirubato/KeyDive/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.8"
coloredlogs = "^15.0.1"
frida = "^16.6.0"
pathlib = "^1.0.1"
pycryptodomex = "^3.21.0"
pywidevine = "^1.8.0"
pathvalidate = "^3.2.1"
requests = "^2.32.3"
xmltodict = "^0.14.2"
Flask = { version = "^3.0.3", optional = true }

[tool.poetry.scripts]
keydive = "keydive.__main__:main"

[[tool.poetry.source]]
name = "localpypi"
url = "https://pypi.org/simple/"
priority = "primary"

[certificates]
localpypi = { cert = false }


